+++ PostHog/App Life Cycle/PostHogAppLifeCycleIntegration.swift
@@ -23,7 +23,10 @@ final class PostHogAppLifeCycleIntegration: PostHogIntegration {
 
     private weak var postHog: PostHogSDK?
 
-    private var appFromBackground = false
+    // True if the app is launched for the first time
+    private var isFreshAppLaunch = true
+    // Manually maintained flag to determine background status of the app
+    private var isAppBackgrounded: Bool = true
 
     private var didBecomeActiveToken: RegistrationToken?
     private var didEnterBackgroundToken: RegistrationToken?
@@ -144,15 +147,22 @@ final class PostHogAppLifeCycleIntegration: PostHogIntegration {
     private func captureAppOpened() {
         guard let postHog else { return }
 
+        guard isAppBackgrounded else {
+            hedgeLog("Skipping Application Opened event - app already in foreground")
+            return
+        }
+
+        isAppBackgrounded = false
+
         if !postHog.config.captureApplicationLifecycleEvents {
             hedgeLog("Skipping Application Opened event - captureApplicationLifecycleEvents is disabled in configuration")
             return
         }
 
         var props: [String: Any] = [:]
-        props["from_background"] = appFromBackground
+        props["from_background"] = !isFreshAppLaunch
 
-        if !appFromBackground {
+        if isFreshAppLaunch {
             let bundle = Bundle.main
 
             let versionName = bundle.infoDictionary?["CFBundleShortVersionString"] as? String
@@ -165,7 +175,7 @@ final class PostHogAppLifeCycleIntegration: PostHogIntegration {
                 props["build"] = versionCode
             }
 
-            appFromBackground = true
+            isFreshAppLaunch = false
         }
 
         postHog.capture("Application Opened", properties: props)
@@ -173,6 +183,13 @@ final class PostHogAppLifeCycleIntegration: PostHogIntegration {
 
     private func captureAppBackgrounded() {
         guard let postHog else { return }
+
+        guard !isAppBackgrounded else {
+            hedgeLog("Skipping Application Opened event - app already in background")
+            return
+        }
+
+        isAppBackgrounded = true
 
         if !postHog.config.captureApplicationLifecycleEvents {
             hedgeLog("Skipping Application Backgrounded event - captureApplicationLifecycleEvents is disabled in configuration")