+++ PostHogTests/PostHogSessionManagerTest.swift
@@ -24,7 +24,7 @@ enum PostHogSessionManagerTest {
         }
 
         @Test("Session id is cleared after 30 min of background time")
-        func testSessionClearedBackgrounded() throws {
+        func sessionClearedBackgrounded() throws {
             let mockNow = MockDate()
             now = { mockNow.date }
 
@@ -39,20 +39,52 @@ enum PostHogSessionManagerTest {
 
             #expect(newSessionId == originalSessionId)
 
-            mockAppLifecycle.simulateAppDidEnterBackground()
+            mockAppLifecycle.simulateAppDidEnterBackground() // user backgrounds app
+
             mockNow.date.addTimeInterval(60 * 30) // +30 minutes (session should not rotate)
-            newSessionId = PostHogSessionManager.shared.getSessionId()
+            newSessionId = PostHogSessionManager.shared.getSessionId() // background activity
 
             #expect(newSessionId == originalSessionId)
 
             mockNow.date.addTimeInterval(60 * 1) // past 30 minutes (session should clear)
-            newSessionId = PostHogSessionManager.shared.getSessionId()
+            newSessionId = PostHogSessionManager.shared.getSessionId() // background activity, session should be cleared
 
             #expect(newSessionId == nil)
         }
 
-        @Test("Session id is rotated after 30 min of inactivity")
-        func testSessionRotatedWhenInactive() throws {
+        @Test("Session id is cleared after 30 min when moving from background to foreground")
+        func sessionClearedWhenMovingBetweenBackgroundAndForeground() throws {
+            let mockNow = MockDate()
+            now = { mockNow.date }
+
+            let originalSessionId = PostHogSessionManager.shared.getNextSessionId()
+
+            try #require(originalSessionId != nil)
+
+            PostHogSessionManager.shared.touchSession()
+            var newSessionId: String?
+
+            newSessionId = PostHogSessionManager.shared.getSessionId()
+
+            #expect(newSessionId == originalSessionId)
+
+            mockAppLifecycle.simulateAppDidEnterBackground() // user backgrounds app
+            mockNow.date.addTimeInterval(60 * 29) // waits 29 mins
+            mockAppLifecycle.simulateAppDidBecomeActive() // user foregrounds app
+            newSessionId = PostHogSessionManager.shared.getSessionId() // should not rotate
+
+            #expect(newSessionId == originalSessionId)
+
+            mockAppLifecycle.simulateAppDidEnterBackground() // user backgrounds app
+            mockNow.date.addTimeInterval(60 * 31) // waits 30+ mins
+            mockAppLifecycle.simulateAppDidBecomeActive() // user foregrounds app
+            newSessionId = PostHogSessionManager.shared.getSessionId() // *should* rotate
+
+            #expect(newSessionId != originalSessionId)
+        }
+
+        @Test("Session id is rotated after 30 min of inactivity when app is foregrounded")
+        func sessionRotatedWhenInactive() throws {
             let mockNow = MockDate()
             now = { mockNow.date }
 
@@ -81,7 +113,7 @@ enum PostHogSessionManagerTest {
         }
 
         @Test("Session id is rotated after max session length is reached")
-        func testSessionRotatedWhenPastMaxSessionLength() throws {
+        func sessionRotatedWhenPastMaxSessionLength() throws {
             let mockNow = MockDate()
             now = { mockNow.date }
 
@@ -164,7 +196,7 @@ enum PostHogSessionManagerTest {
         }
 
         @Test("Clears $session_id after 30 mins of background inactivity")
-        func testSessionClearedAfterBackgroundInactivity() async throws {
+        func sessionClearedAfterBackgroundInactivity() async throws {
             let sut = getSut(flushAt: 2)
             let mockNow = MockDate()
             now = { mockNow.date }
@@ -197,7 +229,7 @@ enum PostHogSessionManagerTest {
         }
 
         @Test("Rotates $session_id after 30 mins of inactivity")
-        func testSessionRotatedAfterInactivity() async throws {
+        func sessionRotatedAfterInactivity() async throws {
             let sut = getSut(flushAt: 2)
             let mockNow = MockDate()
             now = { mockNow.date }
@@ -235,7 +267,7 @@ enum PostHogSessionManagerTest {
         }
 
         @Test("Rotates $session_id after max session length of 24 hours")
-        func testSessionRotatedAfterMaxSessionLength() async throws {
+        func sessionRotatedAfterMaxSessionLength() async throws {
             let sut = getSut(flushAt: 52)
             let mockNow = MockDate()
             var compoundedTime: TimeInterval = 0
@@ -318,7 +350,7 @@ enum PostHogSessionManagerTest {
         }
 
         @Test("ApplicationLifecyclePublisher handles token deallocation correctly")
-        func testApplicationLifecyclePublisherHandlesTokenDeallocationCorrectly() {
+        func applicationLifecyclePublisherHandlesTokenDeallocationCorrectly() {
             let sut = MockApplicationLifecyclePublisher()
 
             var registrations = [
@@ -349,7 +381,7 @@ enum PostHogSessionManagerTest {
         }
 
         @Test("Session id is NOT cleared after 30 min of background time")
-        func testSessionNotClearedBackgrounded() throws {
+        func sessionNotClearedBackgrounded() throws {
             let mockNow = MockDate()
             now = { mockNow.date }
 
@@ -377,7 +409,7 @@ enum PostHogSessionManagerTest {
         }
 
         @Test("Session id is NOT rotated after 30 min of inactivity")
-        func testSessionNotRotatedWhenInactive() throws {
+        func sessionNotRotatedWhenInactive() throws {
             let mockNow = MockDate()
             now = { mockNow.date }
 
@@ -405,7 +437,7 @@ enum PostHogSessionManagerTest {
         }
 
         @Test("Session id is NOT rotated after max session length is reached")
-        func testSessionNotRotatedWhenPastMaxSessionLength() throws {
+        func sessionNotRotatedWhenPastMaxSessionLength() throws {
             let mockNow = MockDate()
             now = { mockNow.date }
 
@@ -440,7 +472,7 @@ enum PostHogSessionManagerTest {
         }
 
         @Test("Session id is NOT cleared when startSession() is called")
-        func testSessionNotRotatedWhenStartSessionCalled() throws {
+        func sessionNotRotatedWhenStartSessionCalled() throws {
             let mockNow = MockDate()
             now = { mockNow.date }
 
@@ -460,7 +492,7 @@ enum PostHogSessionManagerTest {
         }
 
         @Test("Session id is NOT cleared when endSession() is called")
-        func testSessionNotRotatedWhenEndSessionCalled() throws {
+        func sessionNotRotatedWhenEndSessionCalled() throws {
             let mockNow = MockDate()
             now = { mockNow.date }
 
@@ -480,7 +512,7 @@ enum PostHogSessionManagerTest {
         }
 
         @Test("Session id is NOT rotated when resetSession() is called")
-        func testSessionNotRotatedWhenResetSessionCalled() throws {
+        func sessionNotRotatedWhenResetSessionCalled() throws {
             let mockNow = MockDate()
             now = { mockNow.date }